% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.optimisation.R
\name{agent.optimisation}
\alias{agent.optimisation}
\title{Agent Optimisation}
\usage{
agent.optimisation(fields)
}
\description{
This function chooses the subsidy scenario with the highest utility.
Then the parcel soil type is updated.
}
