% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga.R
\name{ens}
\alias{ens}
\title{Existence Need Satisfaction}
\usage{
ens(fields, subsidy)
}
\description{
the SAGA taken method to create a measure to base the agent decisions on.
}
\details{
Not be edited by farmer type as all farmers need to make a living
}
