% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm.emission.money.R
\name{farm.emissions.money}
\alias{farm.emissions.money}
\title{Farm emissions money}
\usage{
farm.emissions.money(fields, ghg.mapping.data)
}
\description{
Adds the emissions and their associated taxes
}
\details{
Hard programmed to have 2 ha packages.
}
