% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm.peatland.scenarios.R
\name{farm.peatland.scenarios}
\alias{farm.peatland.scenarios}
\title{Farm Peatland Scenarios}
\usage{
farm.peatland.scenarios(fields, peatland.scenarios, y, considered.subsidies)
}
\description{
Reads the economic properties of each peatland scenarios.
}
\details{
Relies on Peatland Scenarios Data function. Could be replaced.
}
