% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm.running.money.R
\name{farm.running.money}
\alias{farm.running.money}
\title{Farm Running Money}
\usage{
farm.running.money(fields, expenses.income)
}
\description{
Adds the taxes and subsidies (as of 12-03-2021) with respect to Barley. Silage, Reed Canary Grass and Willow.
Input should be fields, with organic and crop specified.
}
\details{
This considered the current scenario that the farm is in.
}
