% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.peatland.sub.R
\name{fields.peatland.sub}
\alias{fields.peatland.sub}
\title{Fields Peatland Subsidies}
\usage{
fields.peatland.sub(fields)
}
\description{
Checks the suitability of a field parcel for conversion into peatland.
}
\details{
25-06-2021: Current check is whether the parcel was or is peatland or humus soil.
}
