% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govenment.optimisation.R
\name{govenment.optimisation}
\alias{govenment.optimisation}
\title{Government Optimisation}
\usage{
govenment.optimisation(fields.history)
}
\description{
Checks the iterations of the main model to see if the agree
with Finish climate and economic policies. Also optimises
the subsidy scenario that the government would choose.
}
\details{
Gives statistics on land in conservation.
}
