% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greening_rules.R
\name{greening.alt}
\alias{greening.alt}
\title{Greening Alt}
\usage{
greening.alt(fields, row)
}
\description{
Gives a higher insentive, but uses the same structure as the curent
peatland insentives, however each payout has been increased by 100 euro
per field
}
\details{
Output is per ha
}
