% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greening_rules.R
\name{greening.lq}
\alias{greening.lq}
\title{Greening with a Land Quota}
\usage{
greening.lq(fields, row)
}
\description{
Although this is very similar to the BAU above, this allows smaller farms
to be able to afford the transition into peatland. As they have less spare land
to convert.
}
\details{
DON'T THINK I NEED THIS DO IT ON A NATIONAL LEVEL
Greening with Nature Based Solutions

Allows for the inclusion of nature based solutions.

Biodiversity values taken from the national scheme now
Water values taken from the ground water value, average amount of peat in a field and water retention
2 * 10 000 * 1.41 * 0.9 - 0.6 * 1.34 / 4  = 2 Ha of peat volume with 90 - 60\% saturation (difference between mineral and peat) multiplied by the groundwater rate divided by 4 as not all of it will be used
Carbon taken from ETC

If using more detailed data could vary the water payment
}
