% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga.R
\name{opinion.to.results}
\alias{opinion.to.results}
\title{Opinion to results}
\usage{
opinion.to.results(pop, row, param)
}
\description{
Creates the parameters for the SAGA process based on the opinion of the farmer
}
