% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peatland.utility.R
\name{peatland.utility}
\alias{peatland.utility}
\title{Peatland Utility}
\usage{
peatland.utility(fields, pop, row, scenario)
}
\description{
Generates the Peatland Utility.
}
\details{
Very simple formula
}
