% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_and_geog.R
\name{pop_and_geog}
\alias{pop_and_geog}
\title{Population and Geography builder}
\usage{
pop_and_geog(
  farm.soil.data,
  luke_age_dist,
  luke_voc,
  luke_organic_probs,
  av_farm_size_per_type
)
}
\description{
Build the population and geography based on Luke data.
}
