% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga.R
\name{social.action}
\alias{social.action}
\title{Social Action}
\usage{
social.action(fields, pop)
}
\description{
This is the function that controls the agent behavior. The two
functions ens and ut are checked and based on these numbers four actions can happen.
If these actions are chosen the input dataframe "fields" is updated with the farmer utility
for the said peatland scenario. If not the value is se to 0.
}
