% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga.R
\name{strong.link.true}
\alias{strong.link.true}
\title{Strong Link True}
\usage{
strong.link.true(pop, fields, row)
}
\description{
This function compares the fields in terms of similarity via correlation.
It then considers the farms that have a 70\% correlation and checks what the
subsidies and land uses the similar farms have chosen.
}
