% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga.R
\name{weak.link.true}
\alias{weak.link.true}
\title{Strong Link True}
\usage{
weak.link.true(pop, fields, row)
}
\description{
It then considers all of the farms (unlike stong.link.true) that and checks what the
subsidies and land uses the similar farms have chosen.
}
